+++ .gitpod.yml
@@ -39,7 +39,7 @@ ports:
     onOpen: open-browser
     visibility: public
   - name: backend
-    port: 4567
+    port: 4568
     visibility: public
   - name: xray-daemon
     port: 2000
@@ -48,5 +48,5 @@ ports:
     port: 4318
     visibility: public
   - name: nodejs-sidecar
-    port: 4000
+    port: 4567
     visibility: public
