+++ jwt-node-sidecar/package.json
@@ -13,6 +13,7 @@
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "express-http-proxy": "^1.6.3",
+    "http": "^0.0.1-security",
     "http-proxy": "^1.18.1",
     "http-proxy-middleware": "^2.0.6"
   },
